Comply with lint and format
