+++ src/routes/applet/complex_basics/complex_addition/+page.svelte
@@ -38,7 +38,7 @@
   <Vector2D direction={d1.position} length={d1.position.length()} color={PrimeColor.blue} />
   <Vector2D direction={d2.position} length={d2.position.length()} color={PrimeColor.darkGreen} />
   <Vector2D direction={v_result} length={v_result.length()} color={PrimeColor.raspberry} />
-  <Latex2D latex={'z_3'} position={v_result} color={PrimeColor.raspberry} />
+  <Latex2D latex="z_3" position={v_result} color={PrimeColor.raspberry} />
 
   <Vector2D
     direction={v1_dashed}
